update refactoring of app -- completed with Mentor
