+++ client/components/Root.js
@@ -1,22 +1,24 @@
 import React, { PropTypes } from 'react';
-import {Router, Route, browserHistory } from 'react-router';
+import {IndexRoute, Router, Route, browserHistory } from 'react-router';
 import {Provider} from 'react-redux';
 import App from './App';
+import Main from './Main';
 import Login from './Login';
 import Signup from './Signup';
 import Dashboard from './Dashboard';
-import SnippetForm from './SnippetForm';
 import RequireAuth from '../require_auth';
 
 /** Add back RequireAuth to dashboard **/
 const Root = ({store}) =>(
 	<Provider store={store} >
 		<Router history={browserHistory}>
-			<Route path="/" component={App} />
-			<Route path="/login" component={Login} />
-			<Route path="/signup" component={Signup} />
-			<Route path="/dashboard" component={Dashboard} />
-			<Route path="/add-snippet" component={SnippetForm} />
+			<Route path="/" component={App} >
+				<IndexRoute component={Main} />
+				<Route path="/login" component={Login} />
+				<Route path="/signup" component={Signup} />
+				<Route path="/dashboard" component={Dashboard} />
+				<Route path="/newsnippet" component={SnippetForm} />
+			</Route>
 		</Router>
 	</Provider>
 )
