+++ client/components/Sidebar.js
@@ -1,66 +1,59 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import Snippets from '../components/Snippets';
-import SidebarTop from '../components/SidebarTop';
-import SidebarBottom from '../components/SidebarBottom';
-import * as actions from '../actions/user';
-import * as snippetActions from '../actions/snippet';
 import {browserHistory} from 'react-router';
+import Snippets from '../components/Snippets';
+import * as userActions from '../actions/user';
+import * as snippetActions from '../actions/snippet';
+
 
 class Sidebar extends React.Component {	
 	constructor(props) {
 		super(props);
-		console.log(this.props);
-		this.state = {
-			visible: false,
-			snippets: []
-		};
 		this.toggleSidebar= this.toggleSidebar.bind(this);
 		this.logout = this.logout.bind(this);
 		this.addSnippet = this.addSnippet.bind(this);
 	}
 
-	componentWillMount() {
+	componentDidMount() {
 		// fetch the snippets
 		this.props.dispatch(snippetActions.fetchSnippets());
 	}
 
-	componentWillReceiveProps(nextProps) {
-		this.setState({snippets: nextProps.snippets});
-	}
-
 	toggleSidebar(e) {
 		e.preventDefault();
-		if(this.state.visible) {
-			this.setState({visible: false});
-		} else {
-			this.setState({visible: true});
-		}
+		this.props.dispatch(userActions.toggleSidebar());
 	}
 
 	addSnippet(e) {
 		e.preventDefault();
 		// redirects to creating a snippet
-		browserHistory.push('/add-snippet');
+		browserHistory.push('/newsnippet');
 		console.log('create a snippet');
 	}
 
-	logout(e){
-		e.preventDefault();
-		this.props.dispatch(actions.logout());
-		browserHistory.push('/');
-	}
+	// should be handled on the App.js
+	// logout(e){
+	// 	e.preventDefault();
+	// 	this.props.dispatch(actions.logout());
+	// 	browserHistory.push('/');
+	// }
 
 	render() {
+		const snippetsArray = this.props.snippets.map((snippet, index) => {
+			return (
+				<Snippet title={snippet.title} />
+			)
+		})
+
 		return (
 			<div>
 			<button onClick={this.toggleSidebar} className="sidebar-button">Open/Close</button>
 				
-				<div id="sideBar" className={(this.state.visible? "visible " : "invisible ") + "side-menu"}>
+				<div id="sideBar" className={(this.props.visible? "visible " : "invisible ") + "side-menu"}>
 					<div className="top-menu">
 						<input type="search" placeholder="search" />
 					</div>
-					<Snippets data={this.state.snippetArray} />
+					{this.snippetsArray}
 					<div className="bottom-menu">
 					<button onClick={this.logout}> Logout</button> <button>Settings</button> <button onClick={this.addSnippet}>Create Snippet</button>
 					</div>
@@ -73,8 +66,9 @@ class Sidebar extends React.Component {
 
 const mapStateToProps = (state) => {
 	return {
-		snippets: state.snippetReducer
-			}
+		snippets: state.snippetReducer,
+		sidebarVisible: state.userReducer.sidebarVisible
+	}
 }
 
 export default connect(mapStateToProps)(Sidebar);
