--- client/components/SidebarBottom.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-
-export default function SideBarBottom() {
-	return (
-		<div className="bottom-menu">Logout | Settings</div>
-	);
-}
