--- client/components/SidebarTop.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function SideBarTop() {
-	return (
-		<div className="top-menu">
-			<input type="search" placeholder="search" /><button>+ notes</button>
-		</div>
-	);
-}
