+++ client/components/Snippets.js
@@ -1,7 +1,6 @@
 import React from 'react';
 
 export default function Snippets(props) {
-	console.log(props);
 	return (
 		<div className="list-snippets">
 			<div className="individual-snippets">
