+++ client/reducers/index.js
@@ -1,9 +1,7 @@
 import { combineReducers } from 'redux';
-import userReducer from './userReducer';
+import userReducer from './mainReducer';
 import snippetReducer from './snippetReducer';
-import tokenReducer from './tokenReducer';
 
-
-const reducer = combineReducers({ userReducer, snippetReducer, tokenReducer });
+const reducer = combineReducers({ mainReducer, snippetReducer });
 
 export default reducer;
